Edituser

+++ src/main/java/com/framgia/dao/IUserDAO.java
@@ -2,6 +2,6 @@ package com.framgia.dao;
 
 import com.framgia.model.User;
 
-public interface IUserDAO extends IGenericDAO<User, Long> {
+public interface IUserDAO extends IGenericDAO<User, Integer> {
 
 }
