+++ src/main/java/com/framgia/dao/UserDAO.java
@@ -2,7 +2,7 @@ package com.framgia.dao;
 
 import com.framgia.model.User;
 
-public class UserDAO extends GenericDAO<User, Long> implements IUserDAO {
+public class UserDAO extends GenericDAO<User, Integer> implements IUserDAO {
 	public UserDAO() {
 		super(User.class);
 	}
